Docs: use imported scripts in conf.py
